Wrote FBLS.
This function performs the convolution after the points are
folded. This means that the FFA is only run once per period block, but
the each convolution takes 2*N sums per block, so the savings we're as
big as I expected.

I might be able to increase the speed more if I am careful about the
modulus operators, and type-convert the DelT parameter outside of the
loop.